[Rename] Label component subatoms 폴더에서 atoms 폴더로 이동 [Modify] Label.jsx와 Label.stories.js 사용성 개선을 위한 코드 수정
